requirements.txt
pandas>=2.0
numpy>=1.25
PyYAML>=6.0
pytest>=8.0
